Fix a long-standing typo in hts_lrand48.
The return type of the definition was wrong, causing callers
using the function to get incorrect results in some cases,
mainly depending on the platform and compiler options used.

Note: this function is not used by htslib.  It was only added for
reasons of completeness when we were putting in hts_drand48, used
by htslib/ksort.h.

+++ hts_os.c
@@ -49,7 +49,7 @@ HTSLIB_EXPORT
 double hts_drand48(void) { return drand48(); }
 
 HTSLIB_EXPORT
-double hts_lrand48(void) { return lrand48(); }
+long hts_lrand48(void) { return lrand48(); }
 #endif
 
 // // On Windows when using the MSYS or Cygwin terminals, isatty fails
